requirements.txt
google-genai
python-dotenv
Pillow
moviepy
pydub
openai